4 queries updated. [40478654]

+++ src/main/java/com/napier/sem/Country.java
@@ -36,7 +36,7 @@ public class Country {
     /**
      * Country population
      */
-    public int Population;
+    public long Population;
 
     /**
      * Country LifeExpectancy
